Added Aikar's flags as optional parameters
Worked out the issues and set up optional parameters.  The first on allows you to disable JLine for servers on Windows machines without Visual C++ redistributable.  The other two are Aikar's flags with the first being for servers with RAM set up to 10g while the other is for servers with 12g or more set.