Rearranged directories and added files
Rearranged the base structure of the config files and moved/edited them accordingly.  Added more version specific folder structure and sections for configs using Aikar's flags as well as configs using Etil's flags as options.  Updated the paper 1.19.4 latest configs with the newest build and added configs for the older build that was replaced.